finish all

+++ app/Services/PenjualanKendaraanService.php
@@ -1,5 +1,4 @@
 <?php
-
 namespace App\Services;
 
 use Illuminate\Support\Carbon;
@@ -60,14 +59,20 @@ class PenjualanKendaraanService
 
     public function menambahkanMobil(Request $request)
     {
-        $dataMobil = $request->only(['nama_mobil', 'mesin', 'kapasitas_penumpang', 'tipe', 'id_kendaraan', 'status'=>'ready']);
-        return $this->penjualanKendaraanRepository->menambahkanMobil($dataMobil);
+        if ($this->detailKendaraan($request->id_kendaraan)) { //validasi id kendaraan
+            $dataMobil = $request->only(['nama_mobil', 'mesin', 'kapasitas_penumpang', 'tipe', 'id_kendaraan', 'status' => 'ready']);
+            return $this->penjualanKendaraanRepository->menambahkanMobil($dataMobil);
+        }
+        return 'kendaraan tidak valid';
     }
 
     public function menambahkanMotor(Request $request)
     {
-        $dataMotor = $request->only(['nama_motor', 'mesin', 'tipe_suspensi', 'tipe_transmisi', 'id_kendaraan']);
-        return $this->penjualanKendaraanRepository->menambahkanMotor($dataMotor);
+        if ($this->detailKendaraan($request->id_kendaraan)) { //validasi id kendaraan
+            $dataMotor = $request->only(['nama_motor', 'mesin', 'tipe_suspensi', 'tipe_transmisi', 'id_kendaraan']);
+            return $this->penjualanKendaraanRepository->menambahkanMotor($dataMotor);
+        }
+        return 'kendaraan tidak valid';
     }
 
     public function detailKendaraan($id)
@@ -85,55 +90,71 @@ class PenjualanKendaraanService
         return $this->penjualanKendaraanRepository->detailMotor($id);
     }
 
-    public function updateKendaraan(Request $request, $id)
+    public function updateKendaraan(Request $request, $id) //method ini tidak digunakan, hanya pelengkap saja
     {
         $dataKendaraan = $request->only(['tahun_keluaran', 'warna', 'harga']);
-        return $this->penjualanKendaraanRepository->updateKendaraan($dataKendaraan, $id);   
+        return $this->penjualanKendaraanRepository->updateKendaraan($dataKendaraan, $id);
     }
 
     public function updateMobil(Request $request, $id)
     {
-        $dataMobil = $this->detailMobil($id);
-        if (!$dataMobil->status) {
-            $dataMobil = $request->only(['nama_mobil', 'mesin', 'kapasitas_penumpang', 'tipe', 'id_kendaraan']);
-            return $this->penjualanKendaraanRepository->updateMobil($dataMobil, $id);   
-        }else {
+        $dataMobil = $request->only(['nama_mobil', 'mesin', 'kapasitas_penumpang', 'tipe', 'id_kendaraan']);
+
+        $dataMobilDetail = $this->detailMobil($id);
+        if ($dataMobilDetail) { //cek apakah motor ada
+            if (!$dataMobilDetail->status) { //cek apakan mobil sudah terjual
+                if ($this->detailKendaraan($request->id_kendaraan)) { //validasi id kendaraan
+                    return $this->penjualanKendaraanRepository->updateMobil($dataMobil, $id);
+                }
+                return 'kendaraan tidak valid';
+            }
             return 'tidak bisa mengupdate mobil yang sudah terjual';
         }
+        return 'mobil tidak ditemukan';
     }
 
     public function updateMotor(Request $request, $id)
     {
-        $dataMotor = $this->detailMotor($id);
-        if(!$dataMotor->status){
-            $dataMotor = $request->only(['nama_motor', 'mesin', 'tipe_suspensi', 'tipe_transmisi', 'id_kendaraan']);
-            return $this->penjualanKendaraanRepository->updateMotor($dataMotor, $id);   
-        }else{
+        $dataMotor = $request->only(['nama_motor', 'mesin', 'tipe_suspensi', 'tipe_transmisi', 'id_kendaraan']);
+
+        $dataMotorDetail = $this->detailMotor($id);
+        if ($dataMotorDetail) { //cek apakah motor ada
+            if (!$dataMotorDetail->status) { //cek apakah motor sudah terjual
+                if ($this->detailKendaraan($request->id_kendaraan)) { //validasi id kendaraan
+                    return $this->penjualanKendaraanRepository->updateMotor($dataMotor, $id);
+                }
+                return 'kendaraan tidak valid';
+            }
             return 'tidak bisa mengupdate motor yang sudah terjual';
         }
+        return 'motor tidak ditemukan';
     }
 
     public function deleteMobil($id)
     {
-        return $this->penjualanKendaraanRepository->deleteMobil($id);
+        if ($this->detailMobil($id)) { //periksa apakah mobil ada
+            return $this->penjualanKendaraanRepository->deleteMobil($id);
+        }
+        return 'mobil tidak ditemukan';
     }
 
     public function deleteMotor($id)
     {
-        return $this->penjualanKendaraanRepository->deleteMotor($id);
+        if ($this->detailMotor($id)) { //periksa apakah motor ada
+            return $this->penjualanKendaraanRepository->deleteMotor($id);
+        }
+        return 'motor tidak ditemukan';
     }
 
-
-    // untuk ngeset status kendaraan dan tanggal dibeli
     private function generatePenjualanData()
     {
         $now = Carbon::now('Asia/Jakarta');
         $formattedDate = $now->format('d/m/Y');
-    
+
         return [
             'status' => 'terjual',
             'date' => $formattedDate,
         ];
     }
-    
+
 }
